Add utility functions for waiting a specific number of cycles
Add the cycles_wait function that waits a specific number of
cycles. Provide an x86_64 implementation in assembler. The assembler
version is guaranteed not to have any memory accesses in its wait
loop, which may be the case if GCC decides to spill the variable
containing the stop time (which actually happened).